+++ src/app/log/record/utiles/getTwoPointDistance.ts
